refactor(step-10): extract 3 tag-domain tools to tools/tag-tools.ts
Creates src/tools/tag-tools.ts exporting registerTagTools(server),
which registers all 3 tag-global tools verbatim from main.ts:
- bear-list-tags (with its formatTagTree helper)
- bear-rename-tag
- bear-delete-tag

All schemas, descriptions, annotations, and handler bodies preserved.

main.ts is now a thin entry point: create the McpServer, call
registerNoteTools(server) + registerTagTools(server), attach process
error handlers, connect the stdio transport. Dropped from 969 lines
(pre-refactor) to 55 lines. No server.registerTool(...) calls remain
in main.ts.

No circular dependencies (verified with madge). 37 unit tests pass.
